Make the theme really configurable.
We wrote this in in the beginning but never really tested it thoroughly.

It turns out that Pyramid "Asset Specifications" didn't quite work the
way we thought they did.  They aren't relative to the pyramid
application but to python module names.  The changes here make it so
that we can create themes (say, the "fossrit_tahrir_theme" and the
"threebean_tahrir_theme") and distribute them on pypi.  Enabling them is
as simple now as listing their module name in the development.ini.