Do not show all holders when there are too many
Some badges have so many holders that it slows down the loading of the
page (especially when there are several thousand).

This commit sets a limit on the number of holders displayed by default,
with a link to show the same page but with a full list of holders. This
should improve page load speeds for a lot of the badges, while retaining
the full list for those interested.

The hard limit sets the highest number of holders shown (200).
Anything above this limit will instead show a link that says
"... and X other people", where X = times_awarded - soft_limit.

You might be thinking that it should be X = times_awarded - hard_limit.
The soft limit is in place for a reason. Suppose there are 201 holders.
It would look silly if the page showed 200 holders and a link that says
"... and 1 other person". Instead, we will show the soft limit (150
holders) and a link that says "... and 51 other people".