Add series information to badge page
Add a box to the badge detail page that shows all badges in the same
series. The relationship is determined based on badge name. Current
badge is highlighted with blue-tinted background.

The assumption on name is that all badges in a series will have a
parenthesised series name and a number in the name. The number can be
either in Arabic or Roman numerals.

This holds for almost all current badges in Fedora. The only exceptions
are:

 * Free the Fedora: has numbers, but not in parentheses
 * Top: Gold, Silver and Bronze are not tied the the others; badges with
   numbers actually work in the opposite direction (Top 10 is after Top
   500 in the series)
 * Editor and Corporate Ant series have no indication that the badges
   belong together